Add custom stg templates as an option (#84)
* Added option to add custom templates runtime.


* Improved help print


* Apply suggestions from code review



* Fix build.


* Refs #13762. Fix resource path on windows.


* Refs #13762. Improve console output.


---------

+++ src/main/java/com/eprosima/fastdds/fastddsgen.java
@@ -42,9 +42,13 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Vector;
+import java.util.Map;
+import java.util.HashMap;
 import org.antlr.stringtemplate.StringTemplate;
 import org.antlr.stringtemplate.StringTemplateErrorListener;
 import org.antlr.stringtemplate.StringTemplateGroup;
@@ -84,6 +88,9 @@ public class fastddsgen
     private boolean m_atLeastOneStructure = false;
     protected static String m_localAppProduct = "fastrtps";
     private ArrayList<String> m_includePaths = new ArrayList<String>();
+
+    // Mapping where the key holds the path to the template file and the value the wanted output file name
+    private Map<String, String> m_customStgOutput = new HashMap<String, String>();
 
     private static VSConfiguration m_vsconfigurations[] = {
         new VSConfiguration("Debug DLL", "Win32", true, true),
@@ -213,6 +220,17 @@ public class fastddsgen
                     throw new BadArgumentException("No URL specified after -ppPath argument");
                 }
             }
+            else if (arg.equals("-extrastg"))
+            {
+                if (count + 1 < args.length)
+                {
+                    m_customStgOutput.put(args[count++], args[count++]);
+                }
+                else
+                {
+                    throw new BadArgumentException("Missing arguments for -extrastg");
+                }
+            } 
             else if (arg.equals("-ppDisable"))
             {
                 m_ppDisable = true;
@@ -364,9 +382,25 @@ public class fastddsgen
                             getVersion(), m_publishercode, m_subscribercode);
 
             // Load string templates
-            System.out.println("Loading templates...");
-            TemplateManager.setGroupLoaderDirectories(
-                "com/eprosima/fastdds/idl/templates:com/eprosima/fastcdr/idl/templates");
+            System.out.println("Loading templates from " + System.getProperty("java.class.path"));           
+            // Add path of custom templates to manager search path
+            String extraPaths = "";
+            for (Map.Entry<String, String> entry : m_customStgOutput.entrySet())
+            {
+                Path path = Paths.get(entry.getKey()).getParent();
+                if (path != null)
+                {
+                    extraPaths += ":" + path.toString().replace("\\", "/");
+                }
+                else
+                {
+                    extraPaths += ":./";
+                }
+            }
+        	
+            String templatePaths = "com/eprosima/fastdds/idl/templates:com/eprosima/fastcdr/idl/templates" + extraPaths;
+            System.out.println("Template resource folders: " + templatePaths);
+            TemplateManager.setGroupLoaderDirectories(templatePaths);
 
             // In local for all products
             if (m_os.contains("Windows"))
@@ -402,7 +436,7 @@ public class fastddsgen
             ArrayList<String> includedIDL = new ArrayList<String>();
             for (int count = 0; returnedValue && (count < m_idlFiles.size()); ++count)
             {
-                Project project = process(m_idlFiles.get(count));
+                Project project = process(m_idlFiles.get(count), true);
 
                 for (String include : project.getIDLIncludeFiles())
                 {
@@ -427,7 +461,7 @@ public class fastddsgen
             // Add include idl files
             for (String included : includedIDL)
             {
-                Project inner = process(included);
+                Project inner = process(included, false);
                 if (inner != null && !solution.existsProject(inner.getFile()))
                 {
                     System.out.println("Adding project: " + inner.getFile());
@@ -526,6 +560,7 @@ public class fastddsgen
         System.out.println("\t\t-replace: replaces existing generated files.");
         System.out.println("\t\t-ppDisable: disables the preprocessor.");
         System.out.println("\t\t-ppPath: specifies the preprocessor path.");
+        System.out.println("\t\t-extrastg <template file> <output file name>: specifies a custom template, template location must be in classpath.");
         System.out.println("\t\t-typeros2: generates type naming compatible with ROS2.");
         System.out.println("\t\t-I <path>: add directory to preprocessor include paths.");
         System.out.println("\t\t-d <path>: sets an output directory for generated files.");
@@ -572,7 +607,8 @@ public class fastddsgen
     }
 
     private Project process(
-            String idlFilename)
+            String idlFilename,
+            boolean processCustomTemplates)
     {
         Project project = null;
         System.out.println("Processing the file " + idlFilename + "...");
@@ -580,7 +616,7 @@ public class fastddsgen
         try
         {
             // Protocol CDR
-            project = parseIDL(idlFilename); // TODO: Quitar archivos copiados TypesHeader.stg, TypesSource.stg, PubSubTypeHeader.stg de la carpeta com.eprosima.fastdds.idl.templates
+            project = parseIDL(idlFilename, processCustomTemplates); // TODO: Quitar archivos copiados TypesHeader.stg, TypesSource.stg, PubSubTypeHeader.stg de la carpeta com.eprosima.fastdds.idl.templates
         }
         catch (Exception ioe)
         {
@@ -596,7 +632,8 @@ public class fastddsgen
     }
 
     private Project parseIDL(
-            String idlFilename)
+            String idlFilename,
+            boolean processCustomTemplates)
     {
         boolean returnedValue = false;
         String idlParseFileName = idlFilename;
@@ -697,6 +734,18 @@ public class fastddsgen
                 tmanager.addGroup("TypesSwigInterface");
                 tmanager.addGroup("DDSPubSubTypeSwigInterface");
             }
+            
+            // Load custom templates into manager
+            if (processCustomTemplates)
+            {
+                for (Map.Entry<String, String> entry : m_customStgOutput.entrySet())
+                {
+                    System.out.println("Loading custom template " + entry.getKey() + "...");
+                    Path path = Paths.get(entry.getKey());
+                    String templateName = path.getFileName().toString().substring(0, path.getFileName().toString().lastIndexOf('.'));
+                    tmanager.addGroup(templateName);
+                }
+            }
 
             // Create main template
             TemplateGroup maintemplates = tmanager.createTemplateGroup("main");
@@ -732,11 +781,39 @@ public class fastddsgen
                 // Create information of project for solution
                 project = new Project(ctx.getFilename(), idlFilename, ctx.getDependencies());
 
+                // Create all custom files for template
+                if (processCustomTemplates)
+                {
+                    for (Map.Entry<String, String> entry : m_customStgOutput.entrySet())
+                    {
+                        Path path = Paths.get(entry.getKey());
+                        String templateName = path.getFileName().toString().substring(0, path.getFileName().toString().lastIndexOf('.'));
+                        System.out.println("Generating from custom " + templateName + " to " + entry.getValue());
+
+                        if (returnedValue = Utils.writeFile(m_outputDir + entry.getValue(), maintemplates.getTemplate(templateName), m_replace))
+                        {
+                            // Try to determine if the file is a header file.
+                            if (entry.getValue().contains(".hpp") || entry.getValue().contains(".h"))
+                            {
+                                project.addCommonIncludeFile(entry.getValue());
+                            }
+                            else
+                            {
+                                project.addCommonSrcFile(ctx.getFilename() + entry.getValue());
+                            }
+                        }
+                        else
+                        {
+                            break;
+                        }
+                    }
+                }
+            	
                 System.out.println("Generating Type definition files...");
-                if (returnedValue =
+                if ((returnedValue) && (returnedValue =
                         Utils.writeFile(m_outputDir + ctx.getFilename() + ".h",
                         maintemplates.getTemplate("TypesHeader"),
-                        m_replace))
+                        m_replace)))
                 {
                     if (returnedValue =
                             Utils.writeFile(m_outputDir + ctx.getFilename() + ".cxx",
